updated routes and added layoutComponent
