+++ apps/ip-upload/ip-upload/src/app/app.component.ts
@@ -1,24 +1,7 @@
-import { Component, ChangeDetectionStrategy, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
-import { AuthQuery, User } from '@blockframes/auth';
-import { Observable } from 'rxjs';
+import { Component } from '@angular/core';
 
 @Component({
-  selector: 'ip-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  selector: 'ip-upload-root',
+  template: '<router-outlet></router-outlet>'
 })
-export class AppComponent implements OnInit {
-  public user$: Observable<User>;
-
-  constructor(private auth: AuthQuery, private router: Router) {}
-
-  async ngOnInit() {
-    this.user$ = this.auth.select(state => state.user);
-  }
-
-  public logout() {
-    this.router.navigate([''])
-  }
-}
+export class AppComponent {}
