+++ apps/ip-upload/ip-upload/src/app/app.module.ts
@@ -41,9 +41,10 @@ import { FormComponent } from './form/form.component';
 import { ListComponent } from './list/list.component';
 import { ViewComponent } from './view/view.component';
 import { HomeComponent } from './home/home.component';
+import { LayoutComponent } from './layout/layout.component';
 
 @NgModule({
-  declarations: [AppComponent, FormComponent, ListComponent, ViewComponent, HomeComponent,],
+  declarations: [AppComponent, FormComponent, ListComponent, ViewComponent, HomeComponent, LayoutComponent],
   imports: [
     BrowserModule,
     ReactiveFormsModule,
