+++ apps/ip-upload/ip-upload/src/index.html
@@ -13,7 +13,7 @@
   <meta name="theme-color" content="#1976d2">
 </head>
 <body>
-  <ip-root></ip-root>
+  <ip-upload-root></ip-upload-root>
   <noscript>Please enable JavaScript to continue using this application.</noscript>
 </body>
 </html>
