+++ libs/account/src/lib/account-profile/account-profile.component.html
@@ -21,7 +21,7 @@
           <textarea matInput placeholder="Biography" formControlName="biography"></textarea>
         </mat-form-field>
         <button mat-raised-button color="primary" (click)="updateProfile()">Save</button>
-        <button mat-raised-button color="secondary" routerLink="/account">View profile</button>
+        <button mat-raised-button color="secondary" routerLink="../">View profile</button>
         <button mat-raised-button color="warn" (click)="deleteAccount()">Delete my account</button>
       </form>
     </mat-card-content>
