+++ libs/account/src/lib/account-view/account-view.component.html
@@ -8,7 +8,7 @@
             <span>Personal data</span>
           </mat-panel-title>
           <mat-panel-description>
-              <a routerLink="/account/profile">
+              <a routerLink="./profile">
                 <mat-icon>settings</mat-icon>
               </a>
           </mat-panel-description>
